+++ scripts/macros/monsterFeatures/fiendishSpirit/fiendishSpirit.js
@@ -1,0 +1,6 @@
+import {claws} from './claws.js';
+import {deathThroes} from './deathThroes.js';
+export let fiendishSpirit = {
+    'deathThroes': deathThroes,
+    'claws': claws
+}